+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 12/3-2023
+
+- Pokédex model in the admin page no longer crashes at it attempts to load all entries.
+
 ## 10/03-2023
 
 - Added the world and character model and basic views and templates for viewing and creation (no update views).